use multiple files for drone entity
